[sw-sysemu] [BEMU] Fix clang compilation
Changed some code that used gnu extensions to be standard C++, fixed lots of
printf() formatting arguments, and a couple of subtle bugs that clang flagged.
Now we can compile without warnings both on Linux and macOS; and with gcc or clang.

+++ sw-sysemu/fpu/f32_log2.c
@@ -104,7 +104,7 @@ float32_t f32_log2( float32_t a )
     } else {
         product = ( sigA >> 22 ) ? product : -(product & 0xffffffffffffff80);
     }
-    product = ( exponent << 49 ) | ( 0x1FFFFFFFFFFFFULL & product );
+    product = ( ((uint_fast64_t)exponent) << 49 ) | ( 0x1FFFFFFFFFFFFULL & product );
     /*------------------------------------------------------------------------
     *------------------------------------------------------------------------*/
     shiftDist = softfloat_countLeadingZeros32( product >> 25 );
